Doc tweaks
"inline" and "in line" are different things. Also, expand on the 'delim' description to give some examples (I expect anyone who uses fastread to know what a delimiter is, but I pray for a day when fastread or something like it is just SOP and something new users are exposed to. In the meantime, example-driven docs cost nothing).