Properly ignore embedded nuls.
It looks like the signature of the Token constructor changed, but the
calls in TokenizerLine were not properly updated for the change. It
wasn't causing a compilation error because the hasNull boolean was being
implicitly coerced to an integer, and the 0 was being coerced to a
boolean, (very bad!).

Fixes #338