Update tokenizer.R
Current wording says "Character vector of strings to use for missing values."  To me, this sounds like these will be INSERTED in place of missing values, so I changed the text to "to interpret as missing values". 

Also, would it be worthwhile to caution the user that "NA" strings (North America, Sodium, etc) will be turned into missing values?