Read into a 'melted' structure
We add two new arguments

1. `skip_empty_rows = TRUE`

The `read_*` functions skip empty rows altogether, and still will by default
with this commit.  The new `melt_*` functions in the next commit will *not* skip
empty rows by default, because they can be meaningful when the data is less
regular, e.g. more than one table per file, separated by blank lines.

2. Re-add `guessInteger()` as an option

This was removed from the `read_*()` functions because it would truncate
floating-point values if they weren't encountered until after guess_max.  But
the `melt_*()` functions won't have the same problem and it is useful to know which values are whole numbers and which are not.