Adds much more robust empty line and comment skipping
We now always count empty lines and comments as part of the number of
skips, but also skip any remaining blank lines are comments after the
skips. This allows you to have consistent skip behavior but also makes
it easy to ignore blanks, which is useful with messy inputs.

Fixes #923

Fix test