Additional fixes for quotes within comments / skipped lines
Single quotes used as apostrophes caused the similar problems to #944,
even if the line was skipped.

The skipping logic is now rewritten to hopefully be more clear and
easier to maintain. It is also more robust to these sorts of problems,
we no longer try to ignore content within single quotes, only double quotes.

Fixes #945