Fix regression in writing with duplicated column names
Because of the way https://github.com/tidyverse/readr/pull/1009 was
implemented it caused a bug when writing data with duplicated column
names. It also only preformed the check for `write_delim()`, not for all
the writing functions.

We now do this in a different way, which again restores the ability to
write data with duplicated columns.

Fixes #1169