Check read_builtin()'s inputs (#1361)
* Check read_builtin()'s inputs

This is in reponse to a request from CRAN.

Previously this expectation:

expect_error(read_builtin(AirPassengers, "datasets"))

would call `utils::data()` on malformed input. Eventually that lead to some condition having length greater than 1. So we got an error, but an error that indicates a programming problem.

I think this means `utils::data()` might have a condition length problem, but it's not my problem.

Instead, I'll be more careful to catch obviously bad input in readr itself.

I was able to reproduce what CRAN reported with released R (they used R-devel) with env var settings such as:

_R_CHECK_LENGTH_1_CONDITION_=verbose
_R_CHECK_LENGTH_1_CONDITION_="abort,verbose"

Note that "true" is not enough to surface the issue, because the offending code is inside expect_error().

* Record PR number in NEWS

* cli::cli_abort() requires cli >= 3.0.0