Combine the best parts of `col_select` docs from readr and vroom
As with `[.]name_repair`, the long-term plan would be for these docs to live in vroom and for readr to inherit. But for now, vroom and readr will just have the same docs.

The whole reason I looked at this was to see if I could eliminate this NOTE:

Check: Rd cross-references
Result: NOTE
    Undeclared packages ‘tidyselect’, ‘vctrs’ in Rd xrefs

But I think the link to tidyselect does more good (a link for users) than harm (pedantic CRAN check).