Manual intervention on revdep results
I downloaded the log for MultiATSM from AWS cloud checks and it's clean.

I was also able to check MultiATSM locally.

Not sure why this was reported as a failure by cloud checks. I note that one of MultiATSM's Suggested packages has been archived (neldermead), but that doesn't seem to be the / a problem.