revdepcheck::cloud_check()
ggPMX and xpose.nlmixr have dependencies which require R >= 4.2, but our cloud checks are running on 4.1. Also one of them Suggests a package that's not available from the standard repos.

The NA is tidytags. It has a dependency (graphlayouts) that just updated on CRAN and I and our cloud check machine cannot install from the binary. Using an older binary of graphlayouts allows me to successfully pass tests for tidytags.