Update locale.R example description
Obviously Spanish is one of the many languages in South America, spoken by 9 out of 12 South American countries, but it's still just one of the languages spoken and it's not even the one with the most speakers (which is Portuguese in Brazil, with 212 million inhabitants). 

This change simply makes it explicit that the example, which uses Spanish, is *a* South American locale, and not *the* South American locale.